requirements.txt
h5py==3.6.0
ipympl==0.8.8
ipywidgets==7.6.5
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
librosa==0.9.1
Ninja==1.10.2.3
numpy==1.22.2
pandas==1.4.0
pesq==0.0.4
Pillow==9.0.1
protobuf==3.19.4
pyroomacoustics==0.6.0
pystoi==0.3.3
pytorch-lightning==1.6.5
scipy==1.8.0
sdeint==0.2.4
setuptools==59.5.0  # fixes https://github.com/pytorch/pytorch/issues/69894
seaborn==0.11.2
# torch==1.12.0
torch-ema==0.3
# torchaudio==0.12.0
# torchvision==0.13.0
torchinfo==1.6.3
torchsde==0.2.5
tqdm==4.63.0
wandb==0.12.11
